First end-to-end test
Still using the terrible original workflow, but at least it works on fake data. Next steps are to validate it on real data (need to find a efficient way of doing this, it would be painful to run EM in a test...) and then start changing the API. Also data estimation needs to be added at some point.